src/i2c-dev: force compiler to inline function
